+++ Changes.md
@@ -5,6 +5,7 @@ Fixes
 -----
 
 - LevelSetOffset : Fixed crash when attempting to offset grid types other than `FloatGrid`.
+- AttributeTweaks, CustomAttributes, OptionTweaks, CustomOptions, OptionQuery : Fixed contexts used by "From Scene", "From Selected" and "From Affected" menu items. This fixes errors caused by missing context variables (such as script variables).
 
 1.6.21.4 (relative to 1.6.21.3)
 ========